Fixed compile warning for macos in caffe2ncnn (-Wformat) (#2162)

+++ tools/caffe/caffe2ncnn.cpp
@@ -1527,7 +1527,7 @@ int main(int argc, char** argv)
             }
             else if (bs.dim_size() == 3)
             {
-                fprintf(pp, " 0=%zd 1=%zd 2=-233", size_t(bs.dim(2)), bs.dim(1));
+                fprintf(pp, " 0=%zd 1=%zd 2=-233", size_t(bs.dim(2)), size_t(bs.dim(1)));
             }
             else // bs.dim_size() == 4
             {
